Split out applyPostEdits
